Added test for snapshot store misconfiguration (#18)
* Added test for snapshot store misconfiguration
