+++ src/Akka.HealthCheck.Persistence.Tests/AkkaPersistenceLivenessProbeSubscriptionTest.cs
@@ -5,6 +5,7 @@
 // -----------------------------------------------------------------------
 using Akka.Actor;
 using Akka.HealthCheck.Liveness;
+using Akka.Util.Internal;
 using FluentAssertions;
 using System;
 using Xunit;
@@ -16,27 +17,12 @@ namespace Akka.HealthCheck.Persistence.Tests
     {
 
         public AkkaPersistenceLivenessProbeSubscriptionTest(ITestOutputHelper helper)
-            : base(config, output: helper)
+            : base(TestConfig.GetValidConfigurationString(), output: helper)
         {
 
         }
 
-        public static string config = @"akka.persistence {
-                                         journal {
-                                                    plugin = ""akka.persistence.journal.sqlite""
-                                                    sqlite {
-                                                            class = ""Akka.Persistence.Sqlite.Journal.SqliteJournal, Akka.Persistence.Sqlite""
-                                                            auto-initialize = on
-                                                            connection-string = ""Filename=file:memdb.db;Mode=Memory;Cache=Shared""
-                                                     }}
-                                         snapshot-store {
-                                                plugin = ""akka.persistence.snapshot-store.sqlite""
-                                                sqlite {
-                                                class = ""Akka.Persistence.Sqlite.Snapshot.SqliteSnapshotStore, Akka.Persistence.Sqlite""
-                                                auto-initialize = on
-                                                connection-string = ""Filename=file:memdb.db;Mode=Memory;Cache=Shared""
-                       }
-                   }}";
+
         [Fact(DisplayName = "AkkaPersistenceLivenessProbe should correctly handle subscription requests")]
         public void AkkaPersistenceLivenessProbe_Should_Handle_Subscriptions_In_Any_State()
         {
